==> src/notebooks/1.Limpiar_y_normalizar_datos.ipynb <==
# 1. Limpiar y normalizar datos

import pandas as pd

 ## Attribute Information

* **Age**: age of the patient [years]

* **Sex**: sex of the patient [M: Male, F: Female]

* **ChestPainType**: chest pain type [TA: Typical Angina, ATA: Atypical Angina, NAP: Non-Anginal Pain, ASY: Asymptomatic]

* **RestingBP**: resting blood pressure [mm Hg]

* **Cholesterol**: serum cholesterol [mm/dl]

* **FastingBS**: fasting blood sugar [1: if FastingBS > 120 mg/dl, 0: otherwise]

* **RestingECG**: resting electrocardiogram results [Normal: Normal, ST: having ST-T wave abnormality (T wave inversions and/or ST elevation or depression of > 0.05 mV), LVH: showing probable or definite left ventricular hypertrophy by Estes' criteria]

* **MaxHR**: maximum heart rate achieved [Numeric value between 60 and 202]

* **ExerciseAngina**: exercise-induced angina [Y: Yes, N: No]

* **Oldpeak**: oldpeak = ST [Numeric value measured in depression], ejercicio relativo al descanso (oldpeak). Intervalo de tiempo entre latido y latido.


* **ST_Slope**: the slope of the peak exercise ST segment [Up: upsloping, Flat: flat, Down: downsloping]. La pendiente que sigue el electrocadiograma después de un latido, debe ser plano para considerarse normal.

* **HeartDisease**: output class [1: heart disease, 0: Normal]

* **Link**: [DataSet de Kaggle](https://www.kaggle.com/datasets/fedesoriano/heart-failure-prediction)

# Cargamos el Dataframe:

df = pd.read_csv("heart.csv")
df

# Comprobamos que no hayan ni valores nulos ni duplicados:

df.info()

df.drop_duplicates()

Gracias Dios

# Asignamos valores númericos a las columnas con cadenas de texto:
<h3>

* Asignamos 1 o 0 a las columnas con valores True o False, hombre o mujer...etc, respectivamente.

* Asignamos valores ponderados de forma ascendente a las columnas como "ChestPainType", "RestingECG"... en función de su gravedad e influencia negativa sobre nuestra variable target, "HeartDisease". (A mayor gravedad, mayor será el número asignado)

### SEX 
(Mujer = 0, hombre = 1)

df['Sex'].unique()

df['Sex']=df['Sex'].str.replace('M',"1")
df['Sex']=df['Sex'].str.replace('F',"0")

df['Sex'] = pd.to_numeric(df['Sex'])

df['Sex'].unique()

### ChestPainType

df['ChestPainType'].unique()

df['ChestPainType']=df['ChestPainType'].str.replace('TA',"4")
df['ChestPainType']=df['ChestPainType'].str.replace('ATA',"3")
df['ChestPainType']=df['ChestPainType'].str.replace('ASY',"2")
df['ChestPainType']=df['ChestPainType'].str.replace('NAP',"1")

df['ChestPainType']=df['ChestPainType'].str.replace('A4',"3")
df['ChestPainType'] = pd.to_numeric(df['ChestPainType'])

df['ChestPainType'].unique()

### RestingECG

df["RestingECG"].unique()

df['RestingECG']=df['RestingECG'].str.replace('Normal',"1")
df['RestingECG']=df['RestingECG'].str.replace('ST',"2")
df['RestingECG']=df['RestingECG'].str.replace('LVH',"3")

df['RestingECG'] = pd.to_numeric(df['RestingECG'])


df["RestingECG"].unique()

### ExerciseAngina

df['ExerciseAngina'].unique()

df['ExerciseAngina']=df['ExerciseAngina'].str.replace('N',"0")
df['ExerciseAngina']=df['ExerciseAngina'].str.replace('Y',"1")

df['ExerciseAngina'] = pd.to_numeric(df['ExerciseAngina'])


df['ExerciseAngina'].unique()

### ST_Slope

df['ST_Slope'].unique()

df['ST_Slope']=df['ST_Slope'].str.replace('Up',"2")
df['ST_Slope']=df['ST_Slope'].str.replace('Flat',"1")
df['ST_Slope']=df['ST_Slope'].str.replace('Down',"3")

df['ST_Slope'] = pd.to_numeric(df['ST_Slope'])

* Down = isquemia miocardica 

* Up = posible isquemia miocardica

* Flat = Bien


df['ST_Slope'].unique()

df

# Normalizamos los datos (1,-1)

from sklearn.preprocessing import StandardScaler

ss = StandardScaler()
std_data = ss.fit_transform(df)
origin_data = ss.inverse_transform(std_data)

df.columns

df_std = pd.DataFrame(std_data, columns = ['Age', 'Sex', 'ChestPainType', 'RestingBP', 'Cholesterol', 'FastingBS',
       'RestingECG', 'MaxHR', 'ExerciseAngina', 'Oldpeak', 'ST_Slope',
       'HeartDisease'])

df_normal = pd.DataFrame(origin_data, columns = ['Age', 'Sex', 'ChestPainType', 'RestingBP', 'Cholesterol', 'FastingBS',
       'RestingECG', 'MaxHR', 'ExerciseAngina', 'Oldpeak', 'ST_Slope',
       'HeartDisease'])

df_std

df_normal

# Comprobamos que los datos no se han alterado con la normalización

df.corr()

df_normal.corr()

df_std.corr()

# Exportamos el Dataframe con los datos normalizados: (Para el modelo)

df_std.to_csv("Heart_std.csv", index = False)

# Exportamos el Dataframe limpio: (Para el análisis)

df_normal.to_csv("Heart_limpio.csv", index = False)
