==> src/notebooks/7.Resultados.ipynb <==
## Resultados:

<h2>Usando las siguientes variables:</h2>

**('ST_Slope', 'Oldpeak', 'Cholesterol', 'Age', 'RestingBP', 'MaxHR', 'Sex')**

1. *Modelo_Bagging_Classifier 91,72 %*

2. *Random Forest 87,03 %*

3. *Ada Boost Classifier 86,60 %*

4.  *Decission Tree (Features Importance) 72,67 %*

<h2>Usando las siguientes variables:</h2>

**('Age', 'Sex', 'ChestPainType', 'RestingBP', 'Cholesterol', 'FastingBS',** 
 **'RestingECG', 'MaxHR', 'ExerciseAngina', 'Oldpeak', 'ST_Slope')**
       
1. *Decission Tree 86,33 %*

2. *Regresión Logística 84,74 %*
