==> src/notebooks/8.Export_models.ipynb <==
# Export de todos los modelos:

## Librerías y dataframes:

import pandas as pd
import numpy as np
import os

from sklearn.linear_model import LinearRegression
from sklearn.preprocessing import PolynomialFeatures
from sklearn.linear_model import Ridge
from sklearn.linear_model import LogisticRegression
from sklearn.tree import DecisionTreeRegressor
from sklearn.tree import DecisionTreeClassifier
from sklearn.ensemble import BaggingClassifier
from sklearn.ensemble import RandomForestClassifier
from sklearn.ensemble import RandomForestRegressor
from sklearn.ensemble import AdaBoostClassifier
from sklearn.neighbors import KNeighborsClassifier

from sklearn.preprocessing import MinMaxScaler
from sklearn import metrics
from sklearn.metrics import mean_absolute_error, mean_squared_error, r2_score, mean_absolute_percentage_error
from sklearn.metrics import accuracy_score
from sklearn.model_selection import train_test_split

import joblib

os.chdir("C:/Users/Señor Topo/Desktop/src/data")


df = pd.read_csv("processed/Heart_std.csv")
df2 = pd.read_csv("processed/Heart_limpio.csv")

## 1. Modelo de Regresión Logística:

X = df2[['Age', 'Sex', 'ChestPainType', 'RestingBP', 'Cholesterol', 'FastingBS',
       'RestingECG', 'MaxHR', 'ExerciseAngina', 'Oldpeak', 'ST_Slope']]

y = df2["HeartDisease"]

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.3, random_state=1)

modelo_4 = LogisticRegression(random_state=20)
modelo_4.fit(X_train, y_train);

print("El modelo tiene un porcentaje de acierto del",modelo_4.score(X_test, y_test)*100,"%")

joblib.dump(modelo_4, "4.Modelo_de_Regresión_logística.pkl")

## 2. Árbol de decisiones:

X = df2[['Age', 'Sex', 'ChestPainType', 'RestingBP', 'Cholesterol', 'FastingBS',
       'RestingECG', 'MaxHR', 'ExerciseAngina', 'Oldpeak', 'ST_Slope']]

y = df2["HeartDisease"]

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size = 0.20, random_state = 1)

modelo_5 = DecisionTreeRegressor(max_depth=24, random_state=7)
modelo_5.fit(X_train, y_train)

y_pred = modelo_5.predict(X_test)

print("modelo_2 Accuracy:",metrics.accuracy_score(y_test, y_pred))

print("MAE", mean_absolute_error(y_test, y_pred))
print("MAPE", mean_absolute_percentage_error(y_test, y_pred))
print("MSE", mean_squared_error(y_test, y_pred))
print("RMSE", np.sqrt(mean_squared_error(y_test, y_pred)))

joblib.dump(modelo_5, "5.Modelo_de_Arbol_de_decisiones.pkl")

## 3. Árbol de decisiones: (Features Importance)

X = df2[['ST_Slope', 'Oldpeak', "Cholesterol", 'Age', "RestingBP", 'MaxHR',"Sex"]]

y = df2["HeartDisease"]

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size = 0.30, random_state = 2)

arbol_2 = DecisionTreeRegressor(max_depth=60, random_state = 20)
arbol_2.fit(X_train, y_train)

y_pred = arbol_2.predict(X_test)

print("arbol_2 Accuracy:",metrics.accuracy_score(y_test, y_pred),"%")

print("Coeficiente determinación", r2_score(y_test, y_pred))
print("MAE", mean_absolute_error(y_test, y_pred))
print("MAPE", mean_absolute_percentage_error(y_test, y_pred))
print("MSE", mean_squared_error(y_test, y_pred))
print("RMSE", np.sqrt(mean_squared_error(y_test, y_pred)))

joblib.dump(arbol_2, "10.Modelo_de_Arbol_de_decisiones_2.pkl")

## 4. Bagging Classifier:

X = df2[['ST_Slope', 'Oldpeak', "Cholesterol", 'Age', "RestingBP", 'MaxHR',"Sex"]]

y = df2["HeartDisease"]

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size = 0.30, random_state = 2)

estimator = DecisionTreeClassifier(max_depth=60,random_state=56)

modelo_6 = BaggingClassifier(
    base_estimator = estimator,
    n_estimators=50,
    max_samples=300,
    bootstrap=True,
    max_features = 6,
    random_state=1)


modelo_6.fit(X_train, y_train)
y_pred = modelo_6.predict(X_test)
accuracy_score(y_test, y_pred)

joblib.dump(modelo_6, "6.Modelo_Bagging_Classifier.pkl")

## 5. Random Forest Classifier:

X = df2[['ST_Slope', 'Oldpeak', "Cholesterol", 'Age', "RestingBP", 'MaxHR',"Sex"]]

y = df2["HeartDisease"]

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size = 0.30, random_state = 2)

modelo_7 = RandomForestClassifier(n_estimators=500,
                                 max_leaf_nodes=16,
                                 random_state=42)
modelo_7.fit(X_train, y_train)

y_pred_rf = modelo_7.predict(X_test)

accuracy_score(y_test, y_pred_rf)

joblib.dump(modelo_7, "7.Modelo_Random_Forest_Classifier.pkl")

## 6. Random Forest Regressor:

X = df2[['ST_Slope', 'Oldpeak', "Cholesterol", 'Age', "RestingBP", 'MaxHR',"Sex"]]

y = df2["HeartDisease"]

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size = 0.30, random_state = 2)

modelo_8 = RandomForestRegressor(n_estimators=100,
                                 max_leaf_nodes=2,
                                 random_state=51)
modelo_8.fit(X_train, y_train)

y_pred_reg = modelo_8.predict(X_train)
mean_absolute_error(y_train, y_pred_reg)

joblib.dump(modelo_8, "8.Modelo_Random_Forest_Regressor.pkl")

## 7. Ada Boost Classifier:

X = df2[['ST_Slope', 'Oldpeak', "Cholesterol", 'Age', "RestingBP", 'MaxHR',"Sex"]]

y = df2["HeartDisease"]

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size = 0.30, random_state = 2)

estimator = DecisionTreeClassifier(max_depth=1)

modelo_9 = AdaBoostClassifier(base_estimator = estimator,
                             n_estimators=200,
                             learning_rate=0.5,
                             random_state=42)

modelo_9.fit(X_train, y_train)

y_pred = modelo_9.predict(X_test)
accuracy_score(y_test, y_pred)

joblib.dump(modelo_9, "9.Modelo_Ada_Boost_Classifier.pkl")

## 8. Modelo K Neighbors Classifier:

X = df2[['Age', 'Sex', 'ChestPainType', 'RestingBP', 'Cholesterol', 'FastingBS',
       'RestingECG', 'MaxHR', 'ExerciseAngina', 'Oldpeak', 'ST_Slope']]

y = df2["HeartDisease"]

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size = 0.3, random_state=0)

scaler = MinMaxScaler()
X_train = scaler.fit_transform(X_train)
X_test = scaler.transform(X_test)


modelo_10 = KNeighborsClassifier(n_neighbors=9)
modelo_10.fit(X_train,y_train)
print("El modelo 2 tiene un",round(100*modelo_10.score(X_test,y_test),2),"%","de acierto.")

joblib.dump(modelo_10, "10.Modelo_K_Neighbors_Classifier.pkl")
